parser: calc line and col info indexed by token index
We can then re-use this info to track current column number
(aka indentation level) of each "key:value" pair (map) or "- element"
(list). This significantly cleans up the code and opens it up for
better (more informative) errors (which are still a TODO).

This change quite naturally extends to handling of unindented
lists.