yaml: remove a type hint for elems of lists
Previously, I thought it was a reasonable assumption to have
strongly typed `Value.list` but turns out a non-homogeneous lists
in YAML are valid, and therefore, relax this assumption allowing
for mixed-type lists.

Additionally, when declaratively requesting a slice/array of floats,
allow and promote integers to floats. The opposite direction is
forbidden however.